+++ CHANGES.rst
@@ -1,6 +1,14 @@
 ``zodbpickle`` Changelog
 ========================
 
+Unreleased
+----------
+
+- Removed support for the ``bytes_as_strings`` arguments to pickling APIs:
+  the pickles created when that argument was true might not be unpickled
+  without passing ``encoding='bytes'``, which ZODB couldn't reliably enforce.
+  On Py3k, ZODB will be using ``protocol=3`` pickles anyway.
+
 0.4.4 (2013-06-07)
 ------------------